Avoid using `compileClasspath` to resolve POM dependencies.
Motivation:

SLF4J dependency has been expectedly updated to 2.0.7 in the
[1.24.3-SNAPSHOT](https://oss.sonatype.org/content/repositories/snapshots/com/linecorp/armeria/armeria/1.24.3-SNAPSHOT/armeria-1.24.3-20230814.032459-136.pom)
although SLF4J 1.7.36 version is explcitly set as a dependency.
https://github.com/line/armeria/blob/7d29cdf1c67fd9e23589cace86d54d3a84b5554e/dependencies.toml#L124-L124
https://github.com/line/armeria/blob/7d29cdf1c67fd9e23589cace86d54d3a84b5554e/build.gradle#L202-L202

After digging into the problem, it turned out that Logback 1.4.8 added
as `compileOnly` dependency in #5079 caused to bump up the SLF4J version
in the POM file.

It was weired that `compileOnly` dependencies affected `implementation`
or `api` dependencies. The culprit of the problem was the `versionMapping`
strategy declared in `java-publish.gradle`.
[Resolved versions](https://docs.gradle.org/current/userguide/publishing_maven.html#publishing_maven:resolved_dependencies)
that publishes the versions resolved during the build are used for all
variants.
https://github.com/line/armeria/blob/7d29cdf1c67fd9e23589cace86d54d3a84b5554e/gradle/scripts/lib/java-publish.gradle#L11-L16

Modifications:

- Changed the strategy to resolve depdencies only from `runtimeClasspath`.
  - It does not make sence to use compileOnly dependencies for
    `apiElements`.
  - The configuration is in the same format as shown in the gradle
    documnetation example. https://docs.gradle.org/current/userguide/publishing_maven.html#publishing_maven:resolved_dependencies

Result:

Fixed the wrong SLF4J dependency in the POM file.
Note that #5079 has not been released. So users are not affected by the
regression.