Allow zero TTL for DNS cache
Motivation:

When creating a new connection, the DNS records cache is disabled to
obtain a new IP address, but there may be a need to cache NXDOMAIN.
The default `DnsCache` shares cache for positive and negative results
and does not allow zero TTL, so it is difficult to configure the desired
settings with `DnsCacheBuilder`.

Modifications:

- Allow zero TTL for successful DNS records.

Result:

`DnsCache` now allows zero TTL for the resolved DNS records.