Upload Gradle build scan (#5104)
Motivation:

It must be useful to see test results on a website and share the link
with other folks. Since Gradle enterprise is activated in the Armeria
project, we can upload build scans to https://ge.armeria.dev.

An access token is required to upload build scans. As GitHub secrets are
not exposed to the pull requests from public forks, it is not possible to
directly upload the build scans in a CI build workflow while keeping the
access token securely.

@ribafish suggested a good idea that uploads `build-scan-data` to GitHub
Actions artifacts and then publishes the backup scan data in a sperate
workflow run on the Armeria repository later.

Modifications:

- Added a new step that uploads Gradle build scan data to GitHub Actions
  artifacts to the CI build.
- Redefined the build job name as a deterministic name.
  - A job name is used to fetch the job result from the JOB API.
  - GitHub does not provide a way to get a job ID or a job name in a workflow,
    - https://github.com/orgs/community/discussions/8945
    - https://stackoverflow.com/questions/71593861/how-to-get-github-actions-matrix-job-name-within-job
- Added some useful GitHub links to the build scan.
- Added `gradle-build-scan.yml` for downloading and executing the
  previous build scan.
- Added `get-pr-number-by-sha.ts` to find a pull request number based on
  the given commit SHA that triggered the build.
  - The pull request number is used to comment the scan results into a pull
    request.
- Miscellaneous)
  - Bump `actions/checkout` to v3 from v2
  - Bump `actions/setup-java` to v3 from v2

Result:

You can now see Gradle build scan of a pull request on https://ge.armeria.dev