Change the retention policy of `@UnstableApi` to `CLASS`
Motivation:

The retention policy of the `@UnstableApi` annotation is currently
`SOURCE`, which means `.class` files in our JAR will not be able to tell
bytecode-level analysis tools about whether a certain class or method is
unstable or not. Therefore, we cannot exclude the unstable API elements
from the API backward compatibility checks we'd like to introduce in the
future.

Modifications:

- Change the retention policy from `SOURCE` to `CLASS` so that our
  `.class` files contains enough metadata about the stability of our
  API.

Result:

- We can exclude the unstable API elements from our future API backward
  compatibility tests.