Introduce `ContextPathServicesBuilder` (context path part 3) (#5157)
Motivation:

As a last part of refactoring, I propose that we introduce a separate
`ContextPathServicesBuilder` which contains the actual logic to add a
`ServiceConfig` to a `VirtualHostBuilder`.

Previousy, the logic for updating was as follows:

```
ServerBuilder#service -> virtualHostTemplate#service -> virtualHostBuilder#addServiceConfigSetters
``` 

I propose that we introduce a `ContextPathServicesBuilder` as an
additional layer when adding `ServiceConfigSetters`.

```
ServerBuilder#service -> virtualHostTemplate#service -> ContextPathServicesBuilder#service -> virtualHostBuilder#addServiceConfigSetters
``` 

The actual logic for creating `ServiceConfig`s will be moved to
`ContextPathServicesBuilder`.
By doing so, I believe we can easily add an extension to add arbitrary
paths like the following in the future:

i.e.
```java
ContextPathServicesBuilder<ServerBuilder> contextPath(String... contextPaths) {
  return new ContextPathServicesBuilder<>(this, defaultVirtualHostBuilder, contextPaths);
}
```

While in the process, I realized that `AnnotatedServiceConfigSetters
annotatedService()` is used as the main way for creating annotated
service bindings. For this reason, I've also created a
`ContextPathAnnotatedServiceConfigSetters` and included the class in
this patch.

final version: https://github.com/line/armeria/pull/5037/files

Modifications:

- Introduced a `ContextPathServicesBuilder` which contains logic for
creating `ServiceConfigSetters`
- Migrated logic for creating `ServiceConfigSetters` from
`VirtualHostBuilder` to `ContextPathServicesBuilder`
- Introduced a `ContextPathAnnotatedServiceConfigSetters` which
overrides `AbstractAnnotatedServiceConfigSetters`. This is done so that
the actual logic in `ContextPathServicesBuilder` can use the
`annotatedService()` API

Result:

- Refactoring is done in preparation for introducing the `ContextPath`
API.

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->

---------

Co-authored-by: minux <songmw725@gmail.com>