Fix a bug where `HttpResponse` can't complete with `CompletionStage` (#5190)
Motivation:

`CompletionStage<T> CompletableFuture.minimalCompletionStage()` returns a `MinimalStage` which implements `CompletionFuture` and raises `UnsupportedOperationException` if the methods only in `CompletionFuture` are called.

`HttpResponse.of(CompletionStage<HttpResponse>)` checks if the given value is an instance of `CompletableFuture` to use an optimized path for `CompletableFuture`. It is not working because `MinimalStage` can be set to the method parameter easily.
```java
CompletableFuture<HttpResponse> future = new CompletableFuture<>();
HttpResponse.of(future.minimalCompletionStage());
```

Discord discussion: https://discord.com/channels/1087271586832318494/1087272728177942629/1153277808630566942

Modifications:

- Do not check if a `CompletationStage` is an instance of `CompletableFuture` in `HttpResponse` and `StreamMessage`.

Result:

You no longer see `UnsupportedOperationException` when creating an `HttpResposne` with `CompletationStage`.