Use the desired protocol to get a pending acquisition (#5173)
Motivation:

HTTP/2 connections expected to reuse are created without waiting for the pending acquisition if a request exceeds the max concurrent streams. A `ChannelAcquisitionFuture` is set with the desired session protocol such as `HTTP` or `HTTPS`. However, `H2C` or `H2` could be used to get a pending acquisition because it uses the actual session protocol. That results in a cache miss and creates additional connections.

Modifications:

- Use the desired session protocol to get a pending acquisition in `handlePiggyback`.
- Change `getPendingAcquisition` to check H2(C) if the desired protocol is `HTTP(S)`. Because it is safe to use HTTP/2 connection since `HTTP(S)` implies HTTP/2.

Result:

When a client exceeds max concurrent streams, a new connection is created and correctly reused from the pending requests.