Fix a bug where `CookieClient` does not work with redirects (#5221)
Motivation:

When `CookieClient` is used while enabling `followRedirects` option, `CookieClient` fails to relay the cookies of the redirection request.

`CookieClient` uses `peekHeaders(callback)` to see the response headers. `RedirectingClient` uses `log().whenAvailable(RequestLogProperty.RESPONSE_HEADERS)` to see response headers.  The key difference is that `peekHeaders()` is invoked when HttpResponse is subscribed or aggregated and `RequestLogProperty.RESPONSE_HEADERS` is immediately invoked when the response headers is received by the response decoder.

As `RedirectingClient` aborts the response after checking the response headers via the request log, the response is not subscribed at all which results in `CookieClient` not working.

Modifications:

- Split the redirect response to access the response headers instead of using `RequestLogProperty.RESPONSE_HEADERS`.
- Used `CancelledSubscriptionException` to silently abort redirect responses and not to log the exception because the response is correctly sent by the server.

Result:

`CookieClient` is now correctly working with redirects.