Fix a bug where responseCause is not propagted to CircuitBreakerClient from RetryingClient
Motivation:

When a child log completes, `responseCause` is propagated as the last
properties with responseEndTime even though an
`UnprocessedRequestException` ocurred before response headers. So
`RESPONSE_CAUSE` is not avaiable when callbacks that depend on
`RequestLogProperty.RESPONSE_HEADERS` or
`RequestLogProperty.RESPONSE_TRAILERS` are executed.
Consequently, `CircuitBreakerRule` can't use the cause to determine the
result.

Modifications:

- Set `responseCause` before other response properties if availble.

Result:

- `RequestLog.responseCause()` is correctly propagated to
  `CircuitBreakerRule` when using `CircuitBreakerClient` with `RetryingClient`.