Fix a bug where responseCause is not propagated to `CircuitBreakerRule` from `RetryingClient` (#5242)
Motivation:

When a child log completes, `responseCause` is propagated as the last
property with responseEndTime even if an
`UnprocessedRequestException` occurred before other response properties.

https://github.com/line/armeria/blob/759bf3b079c0a759d60dc2536fefcefb1cd648d3/core/src/main/java/com/linecorp/armeria/common/logging/DefaultRequestLog.java#L661-L661
`RESPONSE_CAUSE` can not be used when the callbacks that depend on
`RequestLogProperty.RESPONSE_HEADERS` or
`RequestLogProperty.RESPONSE_TRAILERS` are executed. 

https://github.com/line/armeria/blob/3a99e6adb00847422f1345d5a7bfa2d7ab1f8e6d/core/src/main/java/com/linecorp/armeria/client/circuitbreaker/CircuitBreakerClient.java#L350-L350
Consequently, `CircuitBreakerRule` can't use the cause to determine the
result.

Modifications:

- Set `responseCause` before other response properties if availble.

Result:

- `RequestLog.responseCause()` is correctly propagated to
`CircuitBreakerRule` when using `CircuitBreakerClient` with
`RetryingClient`.