Publish `ProxyConnectionEvent` after `HAProxyMessage` is written
Motivation:

A `ProxyConnectionEvent` is published as a connection is established
without waiting for the result of `HAProxyMessage`.
The event is rescheduled on the premise that the HAProxy message would
be written and the HttpSessionHandler would enter the pipeline during
the rescheduration.

It has worked well, but it doesn't provide perfect happen-before.
Since `HttpSessionHandler` is added as the first listener of
`connectionPromise`, if the connection is successful in the callback of
`HAProxyHandler`'s `addListener()`, `HttpSessionHandler` will always
exist in the pipeline.

Therefore we don't need to rely on the reschedule to fire a
`ProxyConnectionEvent` or `ProxyConnectException`

Modifications:

- Fire `ProxyConnectionEvent` when `HAProxyMessage` is succesfully
  written.
- Remove unnessary reschedule jobs.

Result:

`ProxyConnectionEvent` is correctly published after `HAProxyMessage`
is successfully written.