Fixed a bug where an idle connection could not be reused and a new connection is created
Motivation:

For HTTP/2, we expect only as many connections as
`ClientFactoryBuilder.maxNumEventLoopsPerEndpoint()` to be created and
reused. This is working as expected when the connection is actively used.

`DefaultEventLoopScheduler` removes `EndpointLoopState`s that are
inactive for 1 minute. A different eventloop may be assigned for new
requests.

Because of this, active connections (idle timeout > 1 minute) allocated
to the old EventLoop` cannot be reused and a new connection is created
instead.

Modifications:

- Set an idle timeout + 5 seconds as the cleanup interval of
  `DefaultEventLoopScheduler`

Result:

HTTP/2 clients no longer allocate more `EventLoop`s than
`ClientFactoryBuilder.maxNumEventLoopsPerEndpoint()`.