Fix a bug where a connection may be not reused when using a retrying client.
Motivation:

Armeria's `HttpChannelPool` is bound to an `EventLoop`. Different
`EventLoop`s have different `HttpChannelPool`s. In other words, in order
to reuse a connection for an `Endpoint`, the same `EventLoop` must be
selected.

When creating a derived client in `RetryingClient`, a new endpoint is
selected for each retry, but since the `EventLoop` of parent is used as
is. That causes that the `Endpoint` can't use the existing connection
pool for multiplexing and makes a new connection created.

Modifications:

- Use `EventLoopScheduler` to return constant `EventLoop`s for the same
  endpoint.
- Allow setting `EndpointGroup` in `ClientRequestContextBuilder` for
  testability.

Result:

- You no longer not see a connection leak when using `RetryingClient`
  with `EndpointGroup`.