Fix a bug where `RequestLog` does not complete when `HttpResponse` is cancelled.
Motivation:

An `HttpResponse` is cancelled if either the `ResponseHeaders` is
`END_OF_STREAM` or the content is expected to be empty.
https://github.com/line/armeria/blob/9ddd6c1b76fdc0fe8943dd417e80773cb92eebdc/core/src/main/java/com/linecorp/armeria/server/HttpResponseSubscriber.java#L66-L67

If `SubscriptionOption.NOTIFY_CANCELLATION` is set when subscribing,
`CancellationSubscripitonException` should be propagated through
`Subscriber.onError()`. However, there was a bug in `FuseableStreamMessage`
where this error was not propagated. `onError()` signal is ignored if
the `StreamMessage` has been canceled.
https://github.com/line/armeria/blob/756c17797f95c1ddc7afbb8038ad046c080ba605/core/src/main/java/com/linecorp/armeria/common/stream/FuseableStreamMessage.java#L297-L300

Although `SubscriptionOption.NOTIFY_CANCELLATION` is set, there is no
guarantee that `CancellationSubscripitonException` will always be
delivered to `Subscriber.onError()`. If a `StreamMessage` has already
published all elements, `Subscription.cancel()` event will be silently
ignored.

Modifications:

- Intercept using `Subscription.cancel()` in `FilteredStreamMessage` and
  complete `FilteredStreamMessage.whenComplete()` with
  `CancelledSubscriptionException` instead of relying on
  `NOTIFY_CANCELLATION` option.
- Fix `FuseableStreamMessage` to notify cancellation event via
  `Subscriber.onError()` when `NOTIFY_CANCELLATION` option is set.

Result:

Fix a bug where `RequestLog` may be not completed when an `HttpResponse`
only has `ResponsesHeaders`.