Introduce xDS module and SotW stream (#5342)
Motivation:

We would like to introduce a xDS integration module.
In order to fetch resources from xDS module, I would like to introduce a
`Stream` object which represents a single persistent connection to a
remote control plane server. The `Stream` will query a
`SubscriberStorage` to check which resources are subscribed and update
the request accordingly.

Once a response is recieved, the type will be used to fetch a
`ResourceParser` which will later parse the response. Afterwards, the
parameters will be sent over to a `ResponseHandler` which will handle
the response and send either an `ackResponse` or `nackResponse` to keep
the feedback loop open. Also, a `resourcesUpdated` can be sent to
indicate the subscribed resources have been changed.

Modifications:

- Introduced a `SotwXdsStream` which represents a persistent connection
to a control plane server
- Introduced a `ResourceParser` which will be responsible for parsing a
response in a format readily sendable in a nack response
- Introduced a `XdsStreamSubscriber` and `SubscriberStorage` which
represent the currently subscribed resources

Result:

- A basic xDS stream implementation is added
- Step 1 of https://github.com/line/armeria/pull/5336 is merged