Make `WebSocketService` interface
Motivation:

`WebsocketService` is a final class, so it is not easy to make another
service run WebSocket and HTTP request at the same path.

`WebSocketService` is changed as a marker interface in #5280. However,
the interface does not have API, the service cannot take avantage of
the protocol detection of the original `WebSocketService`.  Graphql
WebSocket implmetaion copy and paste similar code for WebSocket
detection.

The design of the code here was determined after discussions with
other maintainers based on the idea from
https://github.com/line/armeria/pull/5280#discussion_r1440384188

Modifications:

- Add `WebSocketService` interface and rename the original one to
  `DefaultWebSocketService`
- Add `WebSocketProtocolHandler` that is in charge of:
  - Upgrade HTTP request to WebSocket
  - Decode HTTP request to WebSocket for inbound
  - Encode WebSocket to HTTP response for outbond
- Refactor `DefaultWebSocketService` to implement new methods
  `WebSocketService` and `WebSocketProtocolHandler`
- Modify Webservicebuilder so that a fallback service can be
  configured.
  - The fallback service would be useful when WebSocket and HTTP request
    is served at the same path.

Result:

- You can now set a fallback service to `WebSocketService` via
  `WebSocketServiceBuilder.fallbackService()`
- Refactor `WebsocketService` to enable composion and delegatation.