Add the ability to run a testNg/testStreaming test on its own (#5338)
Motivation:

Currently, `testNg` is run after `shadedTest` is run when the `check`
command is invoked.
However, even if there are updates to the source code, `testNg` won't
attempt to recompile the source code.
This can be easily reproduced by modifying a testNg class, and trying to
run `./gradlew testNg`.

This is because `testNg` currently depends on the classpath of
`shadedTest`. As long as `shadedTest` isn't invoked, the
`shadedTest.classpath` won't be updated.

Modifications:

- Since `copyShadedTestClasses` declares the test classpath as an input,
it is safe to rely on `copyShadedTestClasses` (just like `shadedTest`
does). Modified `testNg`, `testStreaming` to depend on
`copyShadedTestClasses` instead.
- Removed the `finalizedBy` dependency since it doesn't make sense to
have a dependency between `shadedTest` and `testNg`/`testStreaming`.

Result:

- We can run `./gradlew testNg` and receive results correctly

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->