Fix a bug where all healthy endpoints are removed as `HealthCheckedEndpointGroup` is initialized (#5343)
Motivation:

If new `Endpoint`s are updated in `HealthCheckedEndpointGroup`,
`HealthCheckedEndpointGroup` creates a new `HealthCheckContextGroup`.
When the new `HealthCheckContextGroup` is initialized, the old
`HealthCheckContextGroup` is removed for rolling updates.

The removal logic is added as a fallback to
`contextGroup.whenInitialized()`.
https://github.com/line/armeria/blob/3f54be0ce4370b24977994e247fa3816fde25e29/core/src/main/java/com/linecorp/armeria/client/endpoint/healthcheck/HealthCheckedEndpointGroup.java#L175-L187
Since context groups are stored in the order in which they were
inserted, if the callback added first is executed first, the old value
will always be deleted and the latest value will be maintained.

`CompletableFuture` uses a stack structure for callbacks, so the last
callback will be executed first when it completes. If all contexts see
the same future completion event, the old context group could remove the
new context group. For example, `D` context group removes `A` to `C`
context groups, and then `C` context group removes groups until it finds
itself. But there is no `C` in the context group chain. As a result, `C`
removes `D`, the newest and last one.

This situation will rarely occur when:
- An `HealthCheckedEndpointGroup` is not initialized yet.
- The delegate is updated with new endpoints that have the same value as
the previous endpoints, but there are duplicate endpoints. The new
endpoints create a new `HealthCheckContextGroup` that shares all
`HttpHealthChecker`s with the previous one.

Modifications:

- Do not try to remove the old `HealthCheckContextGroup` with a context
group if it was removed before.
- Fix a bug where the reference count of `DefaultHealthCheckerContext`
is incorrectly counted if there are duplicate endpoints.
- Use the endpoints selected by `HealthCheckStrategy` instead of using
the original endpoints updated by the delegate.

Result:

You no longer see `EndpointSelectionTimeoutException` when
`HealthCheckedEndpointGroup` is initialized with duplicate `Endpoint`s.