eureka: allow overriding the healthCheckUrlPath (#5369)
Motivation:

Zipkin has a custom health check service, in addition to the normal
one of type `HealthCheckService`. This is used to create a custom json
response and predated our use of Armeria. This custom one is mounted at
the path "/health" while the default is "/internal/health".

We recently added Eureka registration, but were not able to change the
path component without having to supply the scheme, host and port as
well. To work around this is a bit hacky, and I would like to remove it
as it cause log warnings.

This change is already supported upstream for similar reasons. See
https://github.com/Netflix/eureka/blob/2b09d77534eae703a36f79f99b27b4d6ed66c3f8/eureka-client/src/main/java/com/netflix/appinfo/InstanceInfo.java#L663-L666

Modifications:

- Added a `healthCheckUrlPath` property to `InstanceInfo` and
  corresponding builders, using a custom constructor to avoid
  serializing it to JSON.
- Modified to use this instead of the default path when set.
- Added an integration test.

Result:

- Health check paths are overridable.