Add a build method to `ServerListener` responsible for graceful termination of an external `Executor`. (#5087)
Motivation:

- Related #5031
- Users may create a dedicated `ExecutorService` to handle background
tasks. In such cases, users may want to gracefully shut down the
`ExecutorService` in sync with Armeria's shutdown. To facilitate this,
it would be useful to add an API that builds a `ServerListener`
responsible for graceful termination. This `ServerListener` would be
able to handle the graceful shutdown of the user's custom
`ExecutorService`.

Modifications:

- Add `stoppingWithExecutor` methods to `ServerListenerBuilder`.
- Extend `ShutdownSupport#of(ExecutorService, ...)` methods with
time-limited termination and customized termination logic.
- Add a factory method `ServerListner#withExecutor` for creating a
complete `ServerListener`.
- Add tests for `ServerListener`.

Result:

- Closes #5031
- Users can easily create and utilize a `ServerListener` that
synchronizes the graceful shutdown of their own custom `ExecutorService`
with Armeria's shutdown.