Fix a leak where WebSocket client receives an invalid response
Motivation:

A leak was detected while a CI build was running.
https://github.com/line/armeria/actions/runs/7665407974/job/20891255893?pr=5167#step:11:4166
It was leaked when a WebSocket client sends a request to an invaild path.

```
08:12:09.002 [armeria-common-worker-epoll-2-11] ERROR io.netty.util.ResourceLeakDetector - LEAK: ByteBuf.release() was not called before it's garbage-collected. See https://netty.io/wiki/reference-counted-objects.html for more information.
Recent access records:
	io.netty.handler.codec.ByteToMessageDecoder.channelRead(ByteToMessageDecoder.java:300)
	io.netty.channel.CombinedChannelDuplexHandler.channelRead(CombinedChannelDuplexHandler.java:251)
	io.netty.channel.AbstractChannelHandlerContext.invokeChannelRead(AbstractChannelHandlerContext.java:442)
	io.netty.channel.AbstractChannelHandlerContext.invokeChannelRead(AbstractChannelHandlerContext.java:420)
...
08:12:09.280 [armeria-common-worker-epoll-2-4] WARN  c.l.a.c.AbstractHttpRequestHandler - [id: 0x351c8c7e, L:/127.0.0.1:40692 - R:localhost/127.0.0.1:57821][h1c] Unexpected exception: a request publisher raised an exception
com.linecorp.armeria.client.websocket.WebSocketClientHandshakeException: invalid response headers: [:status=404, content-length=0]
	at com.linecorp.armeria.client.websocket.DefaultWebSocketClient.validateResponseHeaders(DefaultWebSocketClient.java:174)
	at com.linecorp.armeria.client.websocket.DefaultWebSocketClient.lambda$connect$1(DefaultWebSocketClient.java:117)
```

When the invalid response was received, the error handler aborted an old
`HttpResponse` that was no longer valid. As `SplitResponse` was derived
from the original `HttpResponse`, the ownership had been transferred
too.

Modifications:

- Abort `split.body()` instead of the original response.

Result:

Fixes a leak where a WebSocket client receives an invalid response.

Side note:

In addition to the leak, it seems unsuitable to warn
`Unexpected exception: a request publisher raised an exception` when a
WebSocket client receives an invalid response.
We may abort the request with `RequestCancellationException` instead.