Add Fabric8 Kubernetes client integration (#5167)
Motivation:

Fabric8 is one of the popular Kubernetes client implementations. It has
an abstract layer for HTTP and WebSocket protocols.

If Armeria provides a Kubernetes client to better support the cloud
infrastructure, other useful functions such as #4497 will be able to be
implemented based on it.
 
Modifications:

- Add `kubernetes` module to provide Fabric `StandardHttpClient`.
  - `ArmeriaHttpClientFactory` is automatically activated via Java SPI.
- Both HTTP and WebSocket clients have been implemented in compliance
with Reactive Streams specification.
- WebSocket is working over HTTP/1. WebSocket over HTTP/2 is disabled
for compatibility.
  - Forked test suites from the upstream repo.
- Miscellaneous) 
  - Allow `ProxyConfig` to configure proxy headers.
- Allow `WebSocketClient` to configure `HttpHeaders` and
`RequestOptions` when starting a WebSocket session.
 
Result:

You can use the Fabric Kubernetes client on top of the Armeria client.