Eureka: Remove encoded port from VIP address defaults (#5451)
Motivation:

Currently, when armeria registers a server, it appends `:${port}` to
Eureka VIP addresses to. While it makes sense to encode the port into
the Eureka `serviceId`, the VIP address fields are meant to be hostnames
or IPs. Encoding a port into the VIP address makes these unusable for
certain clients like spring-cloud-sleuth who heuristically look for a
port and only if it is absent trigger Eureka lookups.

Modifications:

- I removed the logic that appended `:${port}` to both the `vipAddress`
and `secureVipAddress`
- I changed the test which actually proved the bug was present!

Result:

- Manual integration with zipkin, sleuth is now able to look up its
endpoint in Eureka.

---------

Signed-off-by: Adrian Cole <adrian@tetrate.io>