Refactor OAuth 2.0 modules to easily implement extension grants
Motivation:

Currently, access token grant flow is tightly coupled with
AccessTokenRequest. As a result, we have to implement
`OAuth2XXXGrant`, `OAuth2XXXGrantBuilder` and `XXXTokenRequest`
to implement a new grant flow.

An extension grant such as JWT is slightly different from the exisiting
access token grant flow. The only difference is in body parameters.
It is not pracitial to create three classes and use internal API for
that.

To resolve the problem, I designed new APIs and refactored the existing
OAuth2 module so that users can easily implement a new extention grant.

API design note:

- `OAuth2Request`: Represents an OAuth 2.0 request
  that can be serialized to `HttpRequest`.
- `AccessTokenRequest`: Provides body parameters and client
  authentication to obtain an access token.
  - This API will be used as an extension point to support new grants.
  - The implmentations of `AccessTokenRequest`
    - `ClientCredentialsAccessTokenRequest`
    - `ResourceOwnerPasswordAccessTokenRequest`
    - `JsonWebTokenAccessTokenRequest`
- `ClientAuthentication`: Provides a client authentication for the OAuth
   2.0 requests. `OAuth2Request` has `ClientAuthentication` as an
   optional value.
  - The interface is a drop-in replacement for `ClientAuthorization` which
    was a final class that does not allow customization.
  - Additionally, as per OAuth 2.0 spec, Client Authentication is a
    correct expession.
- `OAuth2ResponseHandler`: Handles OAuth 2.0 responses.
  - This interface would be useful if users want to handle responses
    with a custom logic.
- `TokenOperationRequest`: Provides body parameters for token
  revocation and token introspection.

Modifications:

- TBU